Initial Files
First upload of the files. I have already worked on this project but did
not think of version control so adding it now. Its never too late.

The status of the project is.

User Module is complete. Thanks to the extension.
Bootstrap Theme. Again thanks to the extension
Rights module, And again thanks to the extension.
Baptism Module, Creating and printing of Baptism Certificates.
Marriage Module, Creating and printing of Marriage Certificates.

Lots more things to come...